core/remote/watcher: Extract squashMoves module
  `RemoteWatcher#squashMoves()` was getting really big and could be
  broken up into different methods.
  Putting it in its own file will help us do that in a clean way.
